requirements.txt
Flask
psycopg2-binary
anthropic
openai
mindspore
mindcv
Pillow
numpy
werkzeug